test/webapp/requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
mysql-connector-python==8.0.26
